Create add function and local storage to preserve the added tasks
