+++ src/index.html
@@ -20,7 +20,7 @@
     <!-- here all the content will be added dinamically -->
   </ul>
 
-  <script src="index.js"></script>
+  <script src="index.js" type="module"></script>
 </body>
 
 </html>
